Implement milestone 1
 - Refactor repository interfaces and implement new UI session repository
- Updated PlaylistRepository to return Result types for create and rename operations.
- Changed reorderPlaylistItem to accept Long for newPosition and added compactPlaylistRanks method.
- Enhanced QueueRepository with replaceQueueSnapshot and updatePlaybackCheckpoint methods.
- Introduced UiSessionRepository for managing UI session state.
- Added comprehensive unit tests for database schema, media persistence, playlist management, and queue operations.
- Implemented Room testing dependencies and configured Robolectric for testing.
- Updated documentation to reflect completed tasks for Room schema and repository transactions.

+++ app/src/main/java/com/app/resn8/domain/model/Playlist.kt
@@ -11,6 +11,6 @@ data class Playlist(
 data class PlaylistItem(
     val playlistId: String,
     val mediaId: String,
-    val position: Double,
+    val position: Long,
     val addedAt: Long = System.currentTimeMillis()
 )
